Add files via upload
The centers design the classes for all ages and do not charge a monthly fee. The owners of Pitt Fitness have found that their clients enjoy the concept of paying per class, resulting in a lively revenue stream with this model. 

The managers at Pitt Fitness use a database management system to record their customers, instructors, classes and class reservations

Each instructor is identified by a unique InstructorID, which contains two letters and two numbers. 

Each class is identified by a unique identification number

Each customer is identified by a unique identification number

Each reservation is identified by a unique number that uses the last two digits of the current year followed by a five-digit number that is incremented

Customers can only pay with a credit card. No cash is accepted

You are hired as a database consultant to develop an operational database to facilitate the needs of the company